+++ wakamiti-engine/wakamiti-core/src/main/java/es/iti/wakamiti/core/datatypes/assertion/BinaryStringAssertProvider.java
@@ -11,10 +11,14 @@ package es.iti.wakamiti.core.datatypes.assertion;
 
 
 import java.text.ParseException;
+import java.util.Arrays;
 import java.util.LinkedHashMap;
+import java.util.LinkedList;
 import java.util.Locale;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
+import es.iti.wakamiti.core.backend.ExpressionMatcher;
 import org.hamcrest.Matcher;
 import org.hamcrest.Matchers;
 
@@ -43,9 +47,8 @@ public class BinaryStringAssertProvider extends AbstractAssertProvider {
 
 
     @Override
-    protected LinkedHashMap<String, Pattern> translatedExpressions(Locale locale) {
-        // binary numeric matchers
-        String[] expressions = {
+    protected String[] expressions() {
+        return new String[] {
                 EQUALS,
                 EQUALS_IGNORE_CASE,
                 EQUALS_IGNORE_WHITESPACE,
@@ -65,8 +68,12 @@ public class BinaryStringAssertProvider extends AbstractAssertProvider {
                 NOT_CONTAINS,
                 NOT_CONTAINS_IGNORE_CASE
         };
+    }
+
+    @Override
+    protected LinkedHashMap<String, Pattern> translatedExpressions(Locale locale) {
         LinkedHashMap<String, Pattern> translatedExpressions = new LinkedHashMap<>();
-        for (String expression : expressions) {
+        for (String expression : expressions()) {
             translatedExpressions.put(
                 expression,
                 Pattern.compile(
@@ -79,6 +86,14 @@ public class BinaryStringAssertProvider extends AbstractAssertProvider {
             );
         }
         return translatedExpressions;
+    }
+
+    @Override
+    protected LinkedList<String> regex(Locale locale) {
+        return Arrays.stream(expressions())
+                .map(exp -> ExpressionMatcher.computeRegularExpression(bundle(locale).getString(exp)))
+                .map(exp -> exp.replace(VALUE_WILDCARD, "\"([^\"\\\\]*(\\\\.[^\"\\\\]*)*)\"|'([^'\\\\]*(\\\\.[^'\\\\]*)*)'"))
+                .collect(Collectors.toCollection(LinkedList::new));
     }