+++ wakamiti-engine/wakamiti-core/src/main/java/es/iti/wakamiti/core/datatypes/assertion/UnaryStringAssertProvider.java
@@ -8,13 +8,11 @@
 package es.iti.wakamiti.core.datatypes.assertion;
 
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Locale;
+import java.util.*;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
+import es.iti.wakamiti.core.backend.ExpressionMatcher;
 import org.hamcrest.Matcher;
 import org.hamcrest.Matchers;
 
@@ -31,21 +29,32 @@ public class UnaryStringAssertProvider extends AbstractAssertProvider {
 
 
     @Override
-    protected LinkedHashMap<String, Pattern> translatedExpressions(Locale locale) {
-        String[] expressions = {
-            NULL,
-            EMPTY,
-            NULL_EMPTY,
-            NOT_NULL,
-            NOT_EMPTY,
-            NOT_NULL_EMPTY
+    protected String[] expressions() {
+        return new String[] {
+                NULL,
+                EMPTY,
+                NULL_EMPTY,
+                NOT_NULL,
+                NOT_EMPTY,
+                NOT_NULL_EMPTY
         };
+    }
+
+    @Override
+    protected LinkedHashMap<String, Pattern> translatedExpressions(Locale locale) {
         LinkedHashMap<String, Pattern> translatedExpressions = new LinkedHashMap<>();
-        for (String key : expressions) {
+        for (String key : expressions()) {
             translatedExpressions
                 .put(key, Pattern.compile(translateBundleExpression(locale, key, "")));
         }
         return translatedExpressions;
+    }
+
+    @Override
+    protected LinkedList<String> regex(Locale locale) {
+        return Arrays.stream(expressions())
+                .map(exp -> ExpressionMatcher.computeRegularExpression(bundle(locale).getString(exp)))
+                .collect(Collectors.toCollection(LinkedList::new));
     }